Re-arranging interface classes
