+++ gui/io/__init__.py
@@ -1,0 +1,2 @@
+from CSVIO import *
+__all__ = ['CSVIO']
